修复单线程区间大于2个G时出现的不能下载的问题

+++ Aria/src/main/java/com/arialyy/aria/core/common/AbsFileer.java
@@ -297,7 +297,7 @@ public abstract class AbsFileer<ENTITY extends AbsNormalEntity, TASK_ENTITY exte
   private void handleBreakpoint() {
     long fileLength = mEntity.getFileSize();
     Properties pro = CommonUtil.loadConfig(mConfigFile);
-    int blockSize = (int) (fileLength / mThreadNum);
+    long blockSize = fileLength / mThreadNum;
     int[] recordL = new int[mThreadNum];
     for (int i = 0; i < mThreadNum; i++) {
       recordL[i] = -1;
